requirements.txt
aiogram==2.20
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
Babel==2.9.1
beautifulsoup4==4.11.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
psycopg2-binary==2.9.3
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
reverso-context-api==0.5
soupsieve==2.3.2.post1
urllib3==1.26.9
yarl==1.7.2
